Change build path (for eclipse) to use java 1.8 library to work with 1.2.0.
Changed the element that skips to next byte boundary to not use a string
for this, but a complex element containing a sequence with alignment 8
(bits).

This is required as daffodil 1.2.0 is more strict about alignment, and
about left-over-data situations. Prior versions would ignore attempts to
align past the end of data. Now if you try to align past the end of data
it sets the position to the end limit position, thereby properly using
up all the data.