Use idiv rather than div for some calculations
Daffodil 2.0.0 fixed the behavior of div and idiv to behave correctly.
The div operator previously acted like idiv when the values were
integers, which is incorrect. div is now always floating point division
and idiv is always integer division. This replaces the uses of div with
idiv so that the equations correctly use integer division as they used
to.