Merge fix to conditionalize future use group properly.
This fixes C/D1 conditionalization.

Refactored the future use groups out so that they don't clutter so much.
Renumbered them to match the spec. 

Fixed test case drawn from the Revision C standard - turns out the
example table in the spec is broken. Worked out the fixed bytes and made
test conform to this.

Prior test of Revision C header required tweeking one byte - with the
fixes for the C/D1 conditionalization this was put back to original
bytes.