Determine the MIL-STD-2045 version from the header
The first field of a MIL-STD-2045 defines the version. According to the
spec he valid values are:

  0000 (0)  MIL-STD-2045-47001
  0001 (1)  MIL-STD-2045-47001B
  0010 (2)  MIL-STD-2045-47001C
  0011 (3)  MIL-STD-2045-47001D
  0100 (4)  MIL-STD-2045-47001D w/CHANGE 1
  0101-1110 (5-14) Undefined
  1111 (15) Version Sent Not Implemented

We currently only support C (2) and D1 (4). This updates the schema to
set the versionSpec variable based on the field, and ensure that the
value is either 2 or 4. Other ways an error is thrown. This variable can
no longer be externally set--it always comes from the data.

This also required updating a few tests where the original data came
from an unsupported version, but is still compatable with the C or D1
version.