Changed daffodil version to 2.4.0-SNAPSHOT
Unparsing (round trip) of test test_2045msghdr2 was failing in 
daffodil 2.3.0 release. Correct unparsing here seems to need a fix that
is on the 2.4.0 development branch - likely the fix to issue 
DAFFODIL-2069, though this schema doesn't involve the specific issue of
bitOrder changes that the bug is about, the place where the unparser
test fails is exactly related to suspensions and computed lengths. 

I verified that daffodil 2.4.0 development branch as of git hash
a6dc540fa13a45e119bbfda77ea4e514ccd1a88f, runs all tests correctly.
That git hash is fixes for DAFFODIL-2078 and DAFFODIL-2069

Note that this test worked in release 2.2.0, so something in 2.3.0 broke
it. We can't revert to 2.2.0 as other things are broken then - byte
order of hexBinary is different, as one instance.