Add tests that exercise C and D1 control_release_marking.
Now depends on daffodil 2.5.0 since that has the fix that allows these
new tests to work. 

However, the DFDL schema for mil-std-2045 has not changed. Only the
daffodil-dependency, and new tests were added.

However, given that without daffodil version 2.5.0 this schema had a
major problem, this dependency change warrants a new version of the
schema. So advanced to version 0.0.5.