Fix XML_VMF version C/D1 enum. Use dfdx:alignmentKind='manual'.
An unparser CLI test on large data was failing due to an alignment-related
circular deadlock, so use of dfdlx:alignmentKind='manual' is an attempt to address this.

There is only one place now where any alignment should occur and that is at the end of the header. This can only be done by way of automatic alignmentKind.

Improved documentation around use of the hdrLen variable.

Fixes https://github.com/DFDLSchemas/mil-std-2045/issues/27