Version 1.3.2 - Updates
Update to Daffodil 3.6.0

Updates of paths for schemaLocation attributes.
Specifically the ones that are absolute paths.

Update schema to avoid duplicate element field names.
This allows JSON output to be legal JSON (no repeated field names).

The version number change is needed as the prior commit changed
element names of the futureUse fields.
That's equivalent to an API change in software so a new version is required.

Added test data files illustrating JSON representation.

Added missing doc files to the doc directory.

This format is public so I believe these specs should be here in
the project.

https://github.com/DFDLSchemas/mil-std-2045/issues/57,
https://github.com/DFDLSchemas/mil-std-2045/issues/56