Version 1.3.3 - Update to match DFDL schema style guide
Updated to Daffodil 4.0.0

Updated to latest DFDL schema style standard.
The exception is namespace URIs which are kept the same because this schema is in extensive use
and changing the URIs would be a breaking incompatible change. Those URIs are:

- `xmlns:ms2045="urn:milstd2045DFDL"`
- `xmlns:msms="urn:milstd2045DFDLMessageSize"`

Cleaned up include/imports and consolidated internal schema files into one. Added Usage instructions to README.md, and release notes.

https://github.com/DFDLSchemas/mil-std-2045/issues/66